requirements.txt
pygame-ce
pygame-menu
pygame-ce
opencv-python
